[feat] : 기존의 axios에서 라우팅 하는 것을 useNavigation을 사용하여 라우팅하게 구조를 변경
