+++ src/utils/NavigateComp.tsx
@@ -1,0 +1,9 @@
+import { useNavigate } from "react-router-dom";
+import SetupInterceptors from "./SetupInterceptors";
+export default function NavigateComp() {
+  const navigator = useNavigate();
+
+  SetupInterceptors(navigator);
+
+  return <></>;
+}
